fix warnings (Windows)

+++ src/crypto/StrKey.cpp
@@ -22,9 +22,9 @@ toStrKey(uint8_t ver, ByteSlice const& bin)
     toEncode.insert(toEncode.end(), bin.begin(), bin.end());
 
     uint16_t crc = crc16((char*)toEncode.data(), (int)toEncode.size());
-    toEncode.emplace_back(crc & 0xFF);
+    toEncode.emplace_back(static_cast<uint8_t>(crc & 0xFF));
     crc >>= 8;
-    toEncode.emplace_back(crc & 0xFF);
+    toEncode.emplace_back(static_cast<uint8_t>(crc & 0xFF));
 
     std::string res;
     res = bn::encode_b32(toEncode);
