+++ src/ledger/LedgerManager.h
@@ -106,11 +106,8 @@ class LedgerManager
     // epoch.
     virtual uint64_t getCloseTime() const = 0;
 
-    // Return the fee required to apply a transaction to the current ledger. The
-    // current ledger's baseFee is a 32bit value in stroops, but it is returned
-    // as a 64bit value here to minimize the chance of overflow in a subsequent
-    // arithmetic operation.
-    virtual int64_t getTxFee() const = 0;
+    // Return the fee required to apply a transaction to the current ledger.
+    virtual uint32_t getTxFee() const = 0;
 
     // return the maximum size of a transaction set to apply to the current
     // ledger
